Fixed bug CORE-6437 : gfix cannot set big value for buffers. While the bug is present in code, affected function is not called when gfix set buffers, thus the bug left invisible for a long time.

+++ src/jrd/cch.cpp
@@ -3578,7 +3578,7 @@ static bool expand_buffers(thread_db* tdbb, ULONG number)
 
 		if (!num_in_seg)
 		{
-			const size_t alloc_size = dbb->dbb_page_size * (num_per_seg + 1);
+			const size_t alloc_size = ((size_t)dbb->dbb_page_size) * (num_per_seg + 1);
 			memory = (UCHAR*) bcb->bcb_bufferpool->allocate(alloc_size ALLOC_ARGS);
 			bcb->bcb_memory.push(memory);
 			memory = FB_ALIGN(memory, dbb->dbb_page_size);
